refactor: move `kill` and `playpause` to seperate files
